Refactor/move path to hdi (#2980)
* refactor: move type defs Anchor, Path, Component from hdk to hdi

* chore: fmt

* doc: changelog

* chore: changelog grammer


* fix: expose Path-related types via hdk prelude

* fix: Component & anchor ROOT paths

* build: bump flake lock

* Revert "build: bump flake lock"

This reverts commit 9ea64661d690c084a0d5233d28e02fe52adc06d7.

---------
