+++ crates/hdi/src/hash_path/shard.rs
