+++ crates/hdi/src/prelude.rs
@@ -8,6 +8,11 @@ pub use crate::entry::must_get_valid_record;
 pub use crate::entry_defs;
 pub use crate::flat_op::*;
 pub use crate::hash::*;
+pub use crate::hash_path::anchor::Anchor;
+pub use crate::hash_path::path::Component;
+pub use crate::hash_path::path::Path;
+pub use crate::hash_path::path::TypedPath;
+pub use crate::hash_path::shard::ShardStrategy;
 pub use crate::hdi::*;
 pub use crate::info::dna_info;
 pub use crate::info::zome_info;
