+++ crates/hdk/src/hash_path/anchor.rs
@@ -1,50 +1,19 @@
-use crate::hash_path::path::Component;
-use crate::hash_path::path::Path;
+use super::path::HdkPathExt;
 use crate::prelude::*;
-use holochain_wasmer_guest::*;
+use hdi::hash_path::{
+    anchor::{Anchor, ROOT},
+    path::{Component, Path},
+};
 
-/// This is the root of the [ `Path` ] tree.
-///
-/// Forms the entry point to all anchors so that agents can navigate down the tree from here.
-pub const ROOT: &[u8; 2] = &[0x00, 0x00];
-
-#[derive(PartialEq, SerializedBytes, serde::Serialize, serde::Deserialize, Debug, Clone)]
-/// An anchor can only be 1 or 2 levels deep as "type" and "text".
-///
-/// The second level is optional and the Strings use the standard [ `TryInto` ] for path [ `Component` ] internally.
-///
-/// __Anchors are required to be included in an application's [ `entry_defs` ]__ callback and so implement all the standard methods.
-/// Technically the [ `Anchor` ] entry definition is the [ `Path` ] definition.
-///
-/// e.g. `entry_defs![Anchor::entry_def()]`
-///
-/// The methods implemented on anchor follow the patterns that predate the Path module but `Path::from(&anchor)` is always possible to use the newer APIs.
-pub struct Anchor {
-    pub anchor_type: String,
-    pub anchor_text: Option<String>,
-}
-
-/// Anchors are just a special case of path, so we can move from anchor to path losslessly.
-/// We simply format the anchor structure into a string that works with the path string handling.
-impl From<&Anchor> for Path {
-    fn from(anchor: &Anchor) -> Self {
-        let mut components = vec![
-            Component::new(ROOT.to_vec()),
-            Component::from(anchor.anchor_type.as_bytes().to_vec()),
-        ];
-        if let Some(text) = anchor.anchor_text.as_ref() {
-            components.push(Component::from(text.as_bytes().to_vec()));
-        }
-        components.into()
-    }
+pub trait TryFromPath {
+    fn try_from_path(path: &Path) -> Result<Anchor, WasmError>;
 }
 
 /// Paths are more general than anchors so a path could be represented that is not a valid anchor.
 /// The obvious example would be a path of binary data that is not valid utf-8 strings or a path
 /// that is more than 2 levels deep.
-impl TryFrom<&Path> for Anchor {
-    type Error = WasmError;
-    fn try_from(path: &Path) -> Result<Self, Self::Error> {
+impl TryFromPath for Anchor {
+    fn try_from_path(path: &Path) -> Result<Self, WasmError> {
         let components: Vec<Component> = path.as_ref().to_owned();
         if components.len() == 2 || components.len() == 3 {
             if components[0] == Component::new(ROOT.to_vec()) {
@@ -168,7 +137,7 @@ where
     let hopefully_anchor_tags: Result<Vec<String>, WasmError> = path
         .children_paths()?
         .into_iter()
-        .map(|path| match Anchor::try_from(&path.path) {
+        .map(|path| match Anchor::try_from_path(&path.path) {
             Ok(anchor) => match anchor.anchor_text {
                 Some(text) => Ok(text),
                 None => Err(wasm_error!(WasmErrorInner::Serialize(
@@ -186,46 +155,6 @@ where
 
 #[cfg(test)]
 #[test]
-fn hash_path_root() {
-    assert_eq!(ROOT, &[0_u8, 0]);
-}
-
-#[cfg(test)]
-#[test]
-fn hash_path_anchor_path() {
-    let examples = [
-        (
-            "foo",
-            None,
-            Path::from(vec![
-                Component::from(vec![0, 0]),
-                Component::from(vec![102, 111, 111]),
-            ]),
-        ),
-        (
-            "foo",
-            Some("bar".to_string()),
-            Path::from(vec![
-                Component::from(vec![0, 0]),
-                Component::from(vec![102, 111, 111]),
-                Component::from(vec![98, 97, 114]),
-            ]),
-        ),
-    ];
-    for (atype, text, path) in examples {
-        assert_eq!(
-            path,
-            (&Anchor {
-                anchor_type: atype.to_string(),
-                anchor_text: text,
-            })
-                .into(),
-        );
-    }
-}
-
-#[cfg(test)]
-#[test]
 fn hash_path_anchor_from_path() {
     let path = Path::from(vec![
         Component::from(vec![0, 0]),
@@ -234,7 +163,7 @@ fn hash_path_anchor_from_path() {
     ]);
 
     assert_eq!(
-        Anchor::try_from(&path).unwrap(),
+        Anchor::try_from_path(&path).unwrap(),
         Anchor {
             anchor_type: "foo".into(),
             anchor_text: Some("bar".into()),
