+++ crates/hdk/src/prelude.rs
@@ -32,9 +32,8 @@ pub use crate::hash_path::anchor::anchor;
 pub use crate::hash_path::anchor::list_anchor_addresses;
 pub use crate::hash_path::anchor::list_anchor_tags;
 pub use crate::hash_path::anchor::list_anchor_type_addresses;
-pub use crate::hash_path::anchor::Anchor;
-pub use crate::hash_path::path::Path;
-pub use crate::hash_path::path::TypedPath;
+pub use crate::hash_path::anchor::TryFromPath;
+pub use crate::hash_path::path::HdkPathExt;
 pub use crate::hdk::*;
 pub use crate::info::agent_info;
 pub use crate::info::call_info;
@@ -70,6 +69,11 @@ pub use hdi;
 pub use hdi::map_extern_infallible;
 pub use hdi::op::OpHelper;
 pub use hdi::prelude::app_entry;
+pub use hdi::prelude::Anchor;
+pub use hdi::prelude::Component;
+pub use hdi::prelude::Path;
+pub use hdi::prelude::ShardStrategy;
+pub use hdi::prelude::TypedPath;
 pub use hdk_derive;
 pub use hdk_derive::hdk_dependent_entry_types;
 pub use hdk_derive::hdk_dependent_link_types;
